pyproject.toml
[project]
name = "ai-mas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.23.0",
    "google-generativeai>=0.8.5",
    "langgraph>=0.5.0",
    "tavily-python>=0.7.8",
]
